Refactor/#74 : 배틀 시작, 종료, 포인트 지급 수정, 배틀 생성시 이미지 파일 사용하도록 변경, 배틀 시간 변경

+++ src/main/java/com/poorlex/poorlex/battle/domain/BattleParticipantWithExpenditure.java
@@ -6,5 +6,5 @@ public interface BattleParticipantWithExpenditure {
 
     BattleParticipant getBattleParticipant();
 
-    int getExpenditure();
+    Long getExpenditure();
 }
